Added option for selecting traversal type: 'depth-first' or 'breadth-first' (the default).
Added queue to manage registering of Tasks with the event loop. The type of queue, FIFO or LIFO, is given by the type of traversal, 'breadth-first' or 'depth-first,' respectively (not really BFS or DFS because asynchronous, but is good qualitative description of crawling strategy). The while loop with Crawler.crawl has been refactored to poll the queue (waiting if its empty) and get the first url to create a processing task to register with event loop. The polling times out after five seconds, which gracefully exits the crawler. Urls are added to the queue at launch and during the processing tasks.